server/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coverage==5.1
cryptography==2.9.2
Django==3.0.6
django-polymorphic==2.1.2
idna==2.9
mysqlclient==1.4.6
pkg-resources==0.0.0
pycparser==2.20
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
